Refactor questions-service using preloaded data
