+++ webapp/backend/gameservice/models/question-model.js
@@ -1,9 +1,10 @@
 const mongoose = require('mongoose');
 
 const questionSchema = new mongoose.Schema({
-    title: String,
-    correctAnswer: String,
-    allAnswers: String,
+    category : String,
+    title : String,
+    correctAnswer : String,
+    allAnswers : String,
 });
 
 const Question = mongoose.model('Question', questionSchema);
